+++ r-kvstore-20150101/alibabacloud_r_kvstore20150101/client.py
@@ -9503,6 +9503,8 @@ class Client(OpenApiClient):
     ) -> r_kvstore_20150101_models.ModifyBackupPolicyResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.backup_retention_period):
+            query['BackupRetentionPeriod'] = request.backup_retention_period
         if not UtilClient.is_unset(request.enable_backup_log):
             query['EnableBackupLog'] = request.enable_backup_log
         if not UtilClient.is_unset(request.instance_id):
@@ -9547,6 +9549,8 @@ class Client(OpenApiClient):
     ) -> r_kvstore_20150101_models.ModifyBackupPolicyResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.backup_retention_period):
+            query['BackupRetentionPeriod'] = request.backup_retention_period
         if not UtilClient.is_unset(request.enable_backup_log):
             query['EnableBackupLog'] = request.enable_backup_log
         if not UtilClient.is_unset(request.instance_id):
@@ -12568,6 +12572,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.auto_pay):
             query['AutoPay'] = request.auto_pay
+        if not UtilClient.is_unset(request.auto_renew):
+            query['AutoRenew'] = request.auto_renew
         if not UtilClient.is_unset(request.business_info):
             query['BusinessInfo'] = request.business_info
         if not UtilClient.is_unset(request.capacity):
@@ -12629,6 +12635,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.auto_pay):
             query['AutoPay'] = request.auto_pay
+        if not UtilClient.is_unset(request.auto_renew):
+            query['AutoRenew'] = request.auto_renew
         if not UtilClient.is_unset(request.business_info):
             query['BusinessInfo'] = request.business_info
         if not UtilClient.is_unset(request.capacity):